requirements.txt
pytest==8.3.3
pytest-django==4.9.0
ruff==0.6.9